Bert_CasRel_RE/requirements.txt
pytorch~=1.12.1
cudatoolkit~=11.3.1
tqdm~=4.64.1
transformers~=4.24.0
numpy~=1.21.5
pandas~=1.5.1
matplotlib~=3.5.1